Rename API_AUTH_TOKEN to AUTH_TOKEN
We have renamed the API_AUTH_TOKEN environment variable to AUTH_TOKEN. This change ensures that the environment variable names are different on both the API and the gem side. It allows us to test different values for each side and reproduce the "unauthorized" error on the API side if needed.